use cfg attribute instead of cfg! macro
There are a number of places where we use the cfg! macro today, but
unfortunately though convenient, that pattern does not completely
isolate conditional parts of the build.  This is particularly pertinent
on Windows, which lacks a lot of symbols that ordinarily appear in a
UNIX libc.

This change has been tested on at least:

* Mac OS X
* Windows 10
* Ubuntu 19.10
* illumos

There are no more compilation errors or warnings, and the program in
test/ appears work correctly.